5922 FIX Windows agent: Many WMI queries could lead to integer overflow
Integer values returned by a number of WMI queries were handled incorrectly,
squeezing 64 bit integers into 32 bit representation. This led to sporadical
erroneous results for a number of agent sections / checks. Potentially
affected were at least the checks wmi_cpuload and ps, but basically any checks
using information provided by WMI. The symptoms of this may have been varying
across different checks.

+++ agents/windows/wmiHelper.h
@@ -79,7 +79,7 @@ private:
 };
 
 template <>
-int Variant::get() const;
+long long Variant::get() const;
 template <>
 bool Variant::get() const;
 template <>
